refactor: change the proper name with variable and methods
